add: abstract authorize handler for authorizationgrant update and create
