+++ Authorization/Infrastructure/Requests/SilentCookieLogin/SilentCookieLoginHandler.cs
@@ -1,22 +1,22 @@
 ﻿using Domain;
 using MediatR;
 using System.Net;
+using Infrastructure.Requests.Abstract;
 using Infrastructure.Services;
 using Infrastructure.Services.Abstract;
 using Microsoft.EntityFrameworkCore;
 
 namespace Infrastructure.Requests.SilentCookieLogin;
-public class SilentCookieLoginHandler : IRequestHandler<SilentCookieLoginCommand, SilentCookieLoginResponse>
+public class SilentCookieLoginHandler : AuthorizeHandler, IRequestHandler<SilentCookieLoginCommand, SilentCookieLoginResponse>
 {
   private readonly IdentityContext _identityContext;
-  private readonly IAuthorizationGrantService _authorizationGrantService;
   
   public SilentCookieLoginHandler(
     IdentityContext identityContext,
     IAuthorizationGrantService authorizationGrantService)
+  : base(identityContext, authorizationGrantService)
   {
     _identityContext = identityContext;
-    _authorizationGrantService = authorizationGrantService;
   }
 
   public async Task<SilentCookieLoginResponse> Handle(SilentCookieLoginCommand request, CancellationToken cancellationToken)
@@ -36,11 +36,11 @@ public class SilentCookieLoginHandler : IRequestHandler<SilentCookieLoginCommand
     string code;
     if (authorizationGrant is null)
     {
-      code = await CreateGrant(request, cancellationToken);
+      code = await CreateGrant(request, request.UserId, cancellationToken);
     }
     else
     {
-      code = await UpdateGrant(authorizationGrant, request, cancellationToken);
+      code = await UpdateGrant(request, authorizationGrant, cancellationToken);
     }
 
     await _identityContext.SaveChangesAsync(cancellationToken: cancellationToken);
@@ -49,55 +49,5 @@ public class SilentCookieLoginHandler : IRequestHandler<SilentCookieLoginCommand
     {
       Code = code
     };
-  }
-
-  private async Task<string> UpdateGrant(
-    AuthorizationCodeGrant authorizationCodeGrant,
-    SilentCookieLoginCommand request,
-    CancellationToken cancellationToken)
-  {
-    var result = await _authorizationGrantService.UpdateAuthorizationGrant(
-      new UpdateAuthorizationGrantArguments
-      {
-          AuthorizationCodeGrant = authorizationCodeGrant,
-          Nonce = request.Nonce,
-          Scope = request.Scope,
-          CodeChallenge = request.CodeChallenge,
-          CodeChallengeMethod = request.CodeChallengeMethod,
-          RedirectUri = request.RedirectUri
-      }, cancellationToken);
-
-    return result.Code;
-  }
-
-  private async Task<string> CreateGrant(SilentCookieLoginCommand request, CancellationToken cancellationToken)
-  {
-    var client = await _identityContext
-      .Set<Client>()
-      .Include(c => c.RedirectUris)
-      .SingleAsync(c => c.Id == request.ClientId, cancellationToken: cancellationToken);
-
-    var session = await _identityContext
-      .Set<User>()
-      .Where(u => u.Id == request.UserId)
-      .SelectMany(u => u.Sessions)
-      .Where(s => !s.IsRevoked)
-      .SingleAsync(cancellationToken: cancellationToken);
-
-    var maxAge = string.IsNullOrWhiteSpace(request.MaxAge) ? client.DefaultMaxAge : long.Parse(request.MaxAge);
-    var result = await _authorizationGrantService.CreateAuthorizationGrant(
-      new CreateAuthorizationGrantArguments
-      {
-        Client = client,
-        Session = session,
-        CodeChallenge = request.CodeChallenge,
-        CodeChallengeMethod = request.CodeChallengeMethod,
-        Nonce = request.Nonce,
-        Scope = request.Scope,
-        RedirectUri = request.RedirectUri,
-        MaxAge = maxAge
-      }, cancellationToken);
-
-    return result.Code;
   }
 }
